+++ cmd/gwdump/sqlite.go
@@ -86,11 +86,11 @@ func dbGetGPUStats(duration string) (*map[int64]map[string]string, error) {
 
 	switch duration {
 	case "current":
-		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.Temp') FROM current ORDER BY ts")
+		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.TempCur') FROM current ORDER BY ts")
 	case "hourly":
-		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.Temp') FROM hourly ORDER BY ts")
+		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.TempCur') FROM hourly ORDER BY ts")
 	case "daily":
-		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.Temp') FROM daily ORDER BY ts")
+		rows, err = db.Query("SELECT ts, host, json_extract(data, '$.Gpu.TempCur') FROM daily ORDER BY ts")
 	}
 	if err != nil {
 		log.Printf("gpustats %s query failed: %s\n", duration, err)